+++ spotfinder/spotfinder.cuh
@@ -117,7 +117,7 @@ void call_do_spotfinding_dispersion(dim3 blocks,
                                     int height,
                                     pixel_t max_valid_pixel_value,
                                     PitchedMalloc<uint8_t> *result_strong,
-                                    int min_count = 3,
+                                    uint8_t min_count = 3,
                                     float nsig_b = 6.0f,
                                     float nsig_s = 3.0f);
 
@@ -132,7 +132,7 @@ void call_do_spotfinding_extended(dim3 blocks,
                                   pixel_t max_valid_pixel_value,
                                   PitchedMalloc<uint8_t> *result_strong,
                                   bool do_writeout = false,
-                                  int min_count = 3,
+                                  uint8_t min_count = 3,
                                   float nsig_b = 6.0f,
                                   float nsig_s = 3.0f,
                                   float threshold = 10.0f);
